tock-register-interface: move tests to crate::fields
Move the tests currently contained in crate::registers to
crate::fields. All of these tests only operate on Field and FieldValue
types and do not actually test the register types themselves.
